+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 [4.0.0]
 -------
 - [x] Single options object with default values
-- [ ] Added more docs
+- [x] Allow module-less config
+- [ ] Add more docs
 
 
 [3.3.0] - 2018-04-21