+++ CHANGELOG.md
@@ -1,5 +1,5 @@
-[4.0.0]
--------
+[4.0.0] - 2018-04-29
+--------------------
 - [x] Single options object with default values
 - [x] Allow module-less / subscriber-less config
 - [x] Modules are accessible via `kojo.modules.moduleName`